# context : restore preemptive sched reset when LLAMA_SET_ROWS=0