@slaren 2025-07-25
| Model         | Test   |   t/s master |   t/s gg/sched-reset-preemt |   Speedup |
|:--------------|:-------|-------------:|----------------------------:|----------:|
| llama 7B Q4_0 | pp512  |      5676.37 |                     5614.36 |      0.99 |
| llama 7B Q4_0 | tg128  |       123.60 |                      148.16 |      1.20 |